core/requirements.txt
# Core
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.3.1
django-environ==0.11.2

# Message Queue
celery==5.5.3
kombu==5.5.4
pika==1.3.2

# Database
psycopg2-binary>=2.9.9
redis==7.0.1

# Email
sendgrid==6.11.0
python-http-client==3.3.7

# Push Notifications
pyfcm==1.5.4
python-jose==3.3.0

# Utils
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.8.0
python-multipart==0.0.9

# Monitoring
sentry-sdk==2.3.1
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
freezegun==1.2.2

# Documentation
drf-yasg==1.21.7

# Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
